apply date range to remaining tardies elements
