+++ app/assets/javascripts/school_administrator_dashboard/dashboard_components/tardies_dashboard/SchoolwideTardies.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import _ from 'lodash';
 
 import DashboardHelpers from '../DashboardHelpers';
 import SchoolTardiesDashboard from './SchoolTardiesDashboard';
@@ -12,29 +11,10 @@ class SchoolwideTardies extends React.Component {
     return DashboardHelpers.tardyEventsByDay(studentRecords);
   }
 
-  homeroomTardyEvents() {
-    let homeroomTardyEvents = {};
-    const studentRecords = this.props.dashboardStudents;
-    const studentsByHomeroom = DashboardHelpers.groupByHomeroom(studentRecords);
-    const schoolYearStart = DashboardHelpers.schoolYearStart();
-    Object.keys(studentsByHomeroom).forEach((homeroom) => {
-      homeroomTardyEvents[homeroom] = 0;
-      _.each(studentsByHomeroom[homeroom], (student) => {
-        student.tardies.forEach((tardy) => {
-          if (moment.utc(tardy.occurred_at).isSameOrAfter(schoolYearStart)) {
-            homeroomTardyEvents[homeroom]++;
-          }
-        });
-      });
-    });
-    return homeroomTardyEvents;
-  }
-
   render() {
     return (
         <SchoolTardiesDashboard
           schoolTardyEvents = {this.schoolTardyEvents()}
-          homeroomTardyEvents = {this.homeroomTardyEvents()}
           dashboardStudents = {this.props.dashboardStudents}/>);
   }
 }
